Use absolute paths when handling diffs
When importing/viewing a diff, patchview-vim gets confused if the
affected files reside in different sub-directories. It appears the
problem could be an automatic change of the current working directory in
between the import of the different files' diffs, however, I was unable
to track it down fully.
As a quick (and dirty) work around, this change switches to using
absolute path names rather than relative ones which fixes the issue.